+++ src-docs/src/views/datagrid/datagrid_styling_example.js
@@ -74,6 +74,7 @@ const gridSnippet = `<EuiDataGrid
     // If showStyleSelector={true} from toolbarVisibility, these last two will be superceded by what the user decides.
     fontSize: 'm',
     cellPadding: 'm',
+    footer: 'overline'
   }}
 />
 `;